feat(chat): add `sendPublicChatMessage` server command
Adds the `sendPublicChatMessage` server command which allows plugins to
send a message in the public chat. When a plugin calls this method, its
name is registered to the message metadata so that it is possible to
later identify which plugin sent the given message. Additionaly plugins
may send a custom metadata to be saved in the messages metadata.