feat(server-commands): adds sendPublicChatMessage and sendCustomPublicChatMessage server commands
This PR adds the sendPublicChatMessage and sendCustomPublicChatMessage server commands which allows plugins to send a message in the public chat. When a plugin calls any of this methods, its name is registered to the message metadata so that it is possible to later identify which plugin sent the given message. Additionally plugins may send a custom metadata to be saved in the messages metadata.

sendPublicChatMessage: sends a message in the public chat which is displayed as sent by the client user
sendCustomPublicChatMessage: sends a custom message in the public chat which is not rendered by the BBB client. This message is meant to be custom-rendered by plugins.