fix(extensible-areas): change return type of `contentFunction`
This change is necessary because the components provided by plugins,
when rendered in the core, were left in the DOM Tree. This occurs
because the plugins and the core are two separate ReactJS applications,
each using its own ReactDOM instance. The `contentFunction` must now return
the root element created by the plugin's ReactDOM instance to allow the core
to properly unmount it.