feat(extensible-area): add apps gallery extensible area
Until now, the only way to add an item in the apps gallery was by
injecting a sidekick generic content. But as the concept of apps
expanded beyond sidebar contents, it was necessary to decouple the
gallery apps from the sidebar content(i.e. generic sidekick content).
So this commit adds an extensible area specifically for the apps gallery.
It has only one type, which is 'ENTRY' and corresponds to an apps gallery
entry that won't open a sidebar content when clicked.